Extract TIndexBuildInfo into a separate file
