+++ ydb/core/tx/schemeshard/schemeshard__init.cpp
@@ -1,5 +1,6 @@
 #include "schemeshard__shred_manager.h"
 #include "schemeshard_impl.h"
+#include "schemeshard_index_build_info.h"
 #include "schemeshard_utils.h"  // for PQGroupReserve
 
 #include <ydb/core/protos/s3_settings.pb.h>
@@ -4705,7 +4706,7 @@ struct TSchemeShard::TTxInit : public TTransactionBase<TSchemeShard> {
                         "Init " << stepName << " BuildInfo not found: id#" << id);
                     return;
                 }
-                auto& buildInfo = *buildInfoPtr->Get();
+                auto& buildInfo = *buildInfoPtr->get();
                 if (!buildInfo.IsBroken) {
                     fillBuildInfoSafe(buildInfo, stepName, fill);
                 }
@@ -4719,7 +4720,7 @@ struct TSchemeShard::TTxInit : public TTransactionBase<TSchemeShard> {
                 }
 
                 while (!rowset.EndOfSet()) {
-                    TIndexBuildInfo::TPtr buildInfo = new TIndexBuildInfo();
+                    auto buildInfo = std::make_shared<TIndexBuildInfo>();
                     fillBuildInfoSafe(*buildInfo, "IndexBuild", [&](TIndexBuildInfo& buildInfo) {
                         TIndexBuildInfo::FillFromRow(rowset, &buildInfo);
                     });
