+++ ydb/core/tx/schemeshard/schemeshard__monitoring.cpp
@@ -1,4 +1,5 @@
 #include "schemeshard_impl.h"
+#include "schemeshard_index_build_info.h"
 
 #include <ydb/core/base/tablet_pipecache.h>
 #include <ydb/core/protos/tx_datashard.pb.h>
@@ -920,7 +921,7 @@ private:
                 }
                 return;
             }
-            const auto& info = *indexInfoPtr->Get();
+            const auto& info = *indexInfoPtr->get();
             TAG(TH4) {str << "Fields";}
             PRE () {
                 str << "BuildInfoId: " << info.Id << Endl
@@ -1023,7 +1024,7 @@ private:
                     }
                     for (auto item : info.Shards) {
                         TShardIdx idx = item.first;
-                        const TIndexBuildInfo::TShardStatus& status = item.second;
+                        const TIndexBuildShardStatus& status = item.second;
                         TABLER() {
                             TABLED() {
                                 str << idx;
