+++ ydb/core/tx/schemeshard/schemeshard_build_index__forget.cpp
@@ -34,7 +34,7 @@ public:
                 TStringBuilder() << "Index build process with id <" << BuildId << "> not found"
             );
         }
-        const auto& indexBuildInfo = *indexBuildInfoPtr->Get();
+        const auto& indexBuildInfo = *indexBuildInfoPtr->get();
         if (indexBuildInfo.DomainPathId != domainPathId) {
             return Reply(
                 Ydb::StatusIds::NOT_FOUND,
