+++ ydb/core/tx/schemeshard/schemeshard_build_index__progress.cpp
@@ -1234,7 +1234,7 @@ private:
         return buildInfo.InProgressShards.empty() && buildInfo.ToUploadShards.empty();
     }
 
-    void AddShard(TIndexBuildInfo& buildInfo, const TShardIdx& idx, const TIndexBuildInfo::TShardStatus& status) {
+    void AddShard(TIndexBuildInfo& buildInfo, const TShardIdx& idx, const TIndexBuildShardStatus& status) {
         switch (status.Status) {
             case NKikimrIndexBuilder::EBuildStatus::INVALID:
             case NKikimrIndexBuilder::EBuildStatus::ACCEPTED:
@@ -1927,7 +1927,7 @@ public:
     bool DoExecute(TTransactionContext& txc, const TActorContext& ctx) override {
         const auto* buildInfoPtr = Self->IndexBuilds.FindPtr(BuildId);
         Y_ENSURE(buildInfoPtr);
-        auto& buildInfo = *buildInfoPtr->Get();
+        auto& buildInfo = *buildInfoPtr->get();
 
         LOG_N("TTxBuildProgress: Execute: " << BuildId << " " << buildInfo.State);
         LOG_D("TTxBuildProgress: Execute: " << BuildId << " " << buildInfo.State << " " << buildInfo);
@@ -2336,7 +2336,7 @@ public:
             } else {
                 LOG_D("InitiateShard " << x.ShardIdx);
             }
-            auto [it, emplaced] = buildInfo.Shards.emplace(x.ShardIdx, TIndexBuildInfo::TShardStatus{std::move(shardRange), ""});
+            auto [it, emplaced] = buildInfo.Shards.emplace(x.ShardIdx, TIndexBuildShardStatus{std::move(shardRange), ""});
             Y_ENSURE(emplaced);
             if (!buildInfo.IsValidatingUniqueIndex()) {
                 shardRange.From = std::move(bound);
@@ -2377,7 +2377,7 @@ public:
         if (!buildInfoPtr) {
             return true;
         }
-        auto& buildInfo = *buildInfoPtr->Get();
+        auto& buildInfo = *buildInfoPtr->get();
 
         if (!GotScheduledBilling(buildInfo)) {
             return true;
@@ -2461,7 +2461,7 @@ public:
         if (!buildInfoPtr) {
             return true;
         }
-        auto& buildInfo = *buildInfoPtr->Get();
+        auto& buildInfo = *buildInfoPtr->get();
         LOG_D("TTxReply : PipeRetry"
             << ", TIndexBuildInfo: " << buildInfo
             << ", shardId# " << ShardId
@@ -2516,7 +2516,7 @@ public:
             return true;
         }
 
-        auto& buildInfo = *buildInfoPtr->Get();
+        auto& buildInfo = *buildInfoPtr->get();
         LOG_D("TTxReply : " << TypeName<TEvResponse>()
             << ", TIndexBuildInfo: " << buildInfo
             << ", record: " << ResponseShortDebugString()
@@ -2539,7 +2539,7 @@ public:
             return true;
         }
 
-        TIndexBuildInfo::TShardStatus& shardStatus = buildInfo.Shards.at(shardIdx);
+        TIndexBuildShardStatus& shardStatus = buildInfo.Shards.at(shardIdx);
         auto actualSeqNo = std::pair<ui64, ui64>(Self->Generation(), shardStatus.SeqNoRound);
         auto recordSeqNo = std::pair<ui64, ui64>(record.GetRequestSeqNoGeneration(), record.GetRequestSeqNoRound());
 
@@ -2618,7 +2618,7 @@ public:
         }
     }
 
-    virtual void HandleProgress(TIndexBuildInfo::TShardStatus& shardStatus, TIndexBuildInfo& buildInfo) {
+    virtual void HandleProgress(TIndexBuildShardStatus& shardStatus, TIndexBuildInfo& buildInfo) {
         Y_ENSURE(false, TStringBuilder() << "HandleProgress is unreachable for " << TypeName<TEvResponse>()
             << ", TIndexBuildInfo: " << buildInfo
             << ", shardStatus: " << shardStatus.ToString());
@@ -2793,7 +2793,7 @@ public:
             return true;
         }
 
-        auto& buildInfo = *buildInfoPtr->Get();
+        auto& buildInfo = *buildInfoPtr->get();
         LOG_D("TTxReply : TEvUploadSampleKResponse"
             << ", TIndexBuildInfo: " << buildInfo
             << ", record: " << record.ShortDebugString());
@@ -2849,7 +2849,7 @@ struct TSchemeShard::TIndexBuilder::TTxReplyValidateUniqueIndex: public TTxShard
         const auto& record = Response->Get()->Record;
         TTabletId shardId = TTabletId(record.GetTabletId());
         TShardIdx shardIdx = Self->GetShardIdx(shardId);
-        TIndexBuildInfo::TShardStatus& shardStatus = buildInfo.Shards.at(shardIdx);
+        TIndexBuildShardStatus& shardStatus = buildInfo.Shards.at(shardIdx);
 
         if (const TString& key = record.GetFirstIndexKey()) {
             shardStatus.Range.From = TSerializedCellVec(key);
@@ -2872,7 +2872,7 @@ struct TSchemeShard::TIndexBuilder::TTxReplyFulltextIndex: public TTxShardReply<
         const auto& record = Response->Get()->Record;
         TTabletId shardId = TTabletId(record.GetTabletId());
         TShardIdx shardIdx = Self->GetShardIdx(shardId);
-        TIndexBuildInfo::TShardStatus& shardStatus = buildInfo.Shards.at(shardIdx);
+        TIndexBuildShardStatus& shardStatus = buildInfo.Shards.at(shardIdx);
 
         shardStatus.DocCount = record.GetDocCount();
         shardStatus.TotalDocLength = record.GetTotalDocLength();
@@ -2893,7 +2893,7 @@ struct TSchemeShard::TIndexBuilder::TTxReplyFulltextDict: public TTxShardReply<T
         if (record.GetFirstTokenRows() || record.GetLastTokenRows()) {
             TTabletId shardId = TTabletId(record.GetTabletId());
             TShardIdx shardIdx = Self->GetShardIdx(shardId);
-            TIndexBuildInfo::TShardStatus& shardStatus = buildInfo.Shards.at(shardIdx);
+            TIndexBuildShardStatus& shardStatus = buildInfo.Shards.at(shardIdx);
 
             shardStatus.FirstToken = record.GetFirstToken();
             shardStatus.FirstTokenRows = record.GetFirstTokenRows();
@@ -2911,7 +2911,7 @@ struct TSchemeShard::TIndexBuilder::TTxReplyProgress: public TTxShardReply<TEvDa
     {
     }
 
-    void HandleProgress(TIndexBuildInfo::TShardStatus& shardStatus, TIndexBuildInfo& buildInfo) override {
+    void HandleProgress(TIndexBuildShardStatus& shardStatus, TIndexBuildInfo& buildInfo) override {
         auto& record = Response->Get()->Record;
 
         if (record.HasLastKeyAck()) {
@@ -3301,7 +3301,7 @@ public:
             return true;
         }
 
-        auto& buildInfo = *buildInfoPtr->Get();
+        auto& buildInfo = *buildInfoPtr->get();
         LOG_I("TTxReply : TEvAllocateResult, id# " << BuildId
             << ", txId# " << txId);
         LOG_D("TTxReply : TEvAllocateResult"
