+++ ydb/core/tx/schemeshard/schemeshard_build_index_tx_base.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "schemeshard_impl.h"
+#include "schemeshard_index_build_info.h"
 
 #include <ydb/public/api/protos/ydb_issue_message.pb.h>
 #include <ydb/public/api/protos/ydb_status_codes.pb.h>
