+++ ydb/core/tx/schemeshard/schemeshard_import.cpp
@@ -1,6 +1,7 @@
 #include "schemeshard_import.h"
 
 #include "schemeshard_impl.h"
+#include "schemeshard_index_build_info.h"
 #include "schemeshard_import_getters.h"
 #include "schemeshard_import_helpers.h"
